refactor: move utils functions to its own file
